Move stubs to client test instance.
